requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
xgboost>=2.1.0
streamlit>=1.40.0
plotly>=5.24.0
matplotlib>=3.9.0
seaborn>=0.13.0
joblib>=1.4.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
nba_api>=1.7.0
requests>=2.32.0
